Models/requirements.txt
# Langchain core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

#  Anthropic Integration
langchain-Anthropic

# Google Gemini (PaLM) integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Envoirement Variable Management
python-dotenv

# Machine Learning Utilites
numpy
scikit-learn